Use new metadata endpoints in mw and share code
