+++ .changeset/forty-goats-pretend.md
@@ -1,0 +1,5 @@
+---
+'@alephium/shared': patch
+---
+
+Move token related types from desktop and mobile wallet
