+++ .changeset/popular-boxes-jam.md
@@ -1,0 +1,5 @@
+---
+'@alephium/mobile-wallet': patch
+---
+
+Use new metadata endpoints for optimization
