+++ apps/desktop-wallet/src/components/AssetLogo.tsx
@@ -16,13 +16,12 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { AssetInfo } from '@alephium/shared'
+import { AssetInfo, NFT } from '@alephium/shared'
 import { ALPH } from '@alephium/token-list'
 import { HelpCircle } from 'lucide-react'
 import styled, { css } from 'styled-components'
 
 import AlephiumLogoSVG from '@/images/alephium_logo_monochrome.svg'
-import { NFT } from '@/types/assets'
 
 interface AssetLogoProps {
   assetId: AssetInfo['id']
