+++ apps/desktop-wallet/src/components/NFTThumbnail.tsx
@@ -16,9 +16,8 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
+import { NFT } from '@alephium/shared'
 import styled from 'styled-components'
-
-import { NFT } from '@/types/assets'
 
 interface NFTThumbnailProps {
   nft: NFT
