+++ apps/desktop-wallet/src/modals/TransactionDetailsModal.tsx
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { addApostrophes } from '@alephium/shared'
+import { addApostrophes, NFT } from '@alephium/shared'
 import { partition } from 'lodash'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
@@ -39,7 +39,6 @@ import ModalPortal from '@/modals/ModalPortal'
 import SideModal from '@/modals/SideModal'
 import { selectAddressIds } from '@/storage/addresses/addressesSelectors'
 import { AddressHash } from '@/types/addresses'
-import { NFT } from '@/types/assets'
 import { AddressConfirmedTransaction } from '@/types/transactions'
 import { formatDateForDisplay, openInWebBrowser } from '@/utils/misc'
 import { getTransactionInfo } from '@/utils/transactions'
