+++ apps/desktop-wallet/src/storage/addresses/addressesSelectors.ts
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { Asset, TokenDisplayBalances } from '@alephium/shared'
+import { Asset, NFT, TokenDisplayBalances } from '@alephium/shared'
 import { ALPH } from '@alephium/token-list'
 import { AddressGroup } from '@alephium/walletconnect-provider'
 import { createSelector } from '@reduxjs/toolkit'
@@ -26,7 +26,6 @@ import { addressesAdapter, contactsAdapter } from '@/storage/addresses/addresses
 import { selectAllAssetsInfo, selectAllNFTs, selectNFTIds } from '@/storage/assets/assetsSelectors'
 import { RootState } from '@/storage/store'
 import { Address, AddressHash } from '@/types/addresses'
-import { NFT } from '@/types/assets'
 import { filterAddressesWithoutAssets } from '@/utils/addresses'
 
 export const {
