+++ apps/desktop-wallet/src/storage/assets/assetsActions.ts
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { Asset, TOKENS_QUERY_LIMIT } from '@alephium/shared'
+import { Asset, SyncUnknownTokensInfoResult, TOKENS_QUERY_LIMIT } from '@alephium/shared'
 import { TokenList } from '@alephium/token-list'
 import { explorer } from '@alephium/web3'
 import { createAction, createAsyncThunk } from '@reduxjs/toolkit'
@@ -26,7 +26,6 @@ import posthog from 'posthog-js'
 import client from '@/api/client'
 import { exponentialBackoffFetchRetry } from '@/api/fetchRetry'
 import { RootState } from '@/storage/store'
-import { SyncUnknownTokensInfoResult } from '@/types/assets'
 
 export const loadingStarted = createAction('assets/loadingStarted')
