+++ apps/desktop-wallet/src/storage/assets/nftsSlice.tsx
@@ -16,12 +16,12 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
+import { NFT } from '@alephium/shared'
 import { createSlice, EntityState } from '@reduxjs/toolkit'
 
 import { syncUnknownTokensInfo } from '@/storage/assets/assetsActions'
 import { nftsAdapter } from '@/storage/assets/assetsAdapter'
 import { customNetworkSettingsSaved, networkPresetSwitched } from '@/storage/settings/networkActions'
-import { NFT } from '@/types/assets'
 
 type NFTsState = EntityState<NFT>
