+++ apps/mobile-wallet/src/screens/SendReceive/Send/AssetsScreen/AssetRow.tsx
@@ -16,7 +16,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { Asset, fromHumanReadableAmount, getNumberOfDecimals, toHumanReadableAmount } from '@alephium/shared'
+import { Asset, fromHumanReadableAmount, getNumberOfDecimals, NFT, toHumanReadableAmount } from '@alephium/shared'
 import { ALPH } from '@alephium/token-list'
 import { MIN_UTXO_SET_AMOUNT } from '@alephium/web3'
 import { useRef, useState } from 'react'
@@ -32,7 +32,6 @@ import Button from '~/components/buttons/Button'
 import Checkmark from '~/components/Checkmark'
 import ListItem from '~/components/ListItem'
 import { useSendContext } from '~/contexts/SendContext'
-import { NFT } from '~/types/assets'
 import { isNft } from '~/utils/assets'
 import { ImpactStyle, vibrate } from '~/utils/haptics'
 import { isNumericStringValid } from '~/utils/numbers'
